Adding Login tests.
